add Splash.Suspense
