+++ src/Dummy.js
@@ -1,0 +1,14 @@
+import { useEffect } from 'react';
+
+const Dummy = ({ onUnmount }) => {
+  useEffect(
+    () => () => {
+      onUnmount();
+    },
+    [onUnmount]
+  );
+
+  return null;
+};
+
+export default Dummy;
